Enabled more warnings to catch bugs
I spent several hours tracing down a bug with uninitialized memory, to
avoid such pains in the future, I enabled some more warnings in hopes of
catching more bugs.

-Wconversion -Wunreachable-code -Wuninitialized -Wmaybe-uninitialized
-Wundef

- Fixed new warnings generated by enabling new warnings

+++ src/asmgen/Block.h
@@ -236,7 +236,7 @@ static void block_link(Block* blk, Block* next) {
     ASSERT(blk != next, "Cannot link block to self");
     for (int i = 0; i < MAX_BLOCK_LINK; ++i ) {
         if (blk->next[i] == 0) {
-            blk->next[i] = next - blk;
+            blk->next[i] = (int)(next - blk);
             return;
         }
     }
